All configuration files were moved to the projects folder

+++ run.sh
@@ -1,12 +1,27 @@
 #!/bin/bash
 
+if [ "$1" == "" ]; then
+    echo 'Usage'
+    echo '1) Copy the projects/sample_project folder to projects/project_name'
+    echo '2) Execute the following command from here: ./run project_name'
+    exit 1
+fi
+
 ROOT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd -P )"
 cd $ROOT_DIR
 
-# Configurações
-# Use relative paths
-OUTPUT_DIR="projects/sample_project/"
+PROJECT_DIR="projects/$1"
+OUTPUT_DIR=$PROJECT_DIR"/output"
 SCRIPT_DIR="src/"
+
+# Remove a classe python de configurações do último projeto
+rm $SCRIPT_DIR"/config.py"
+# Copia a classe python de configurações do projecto para a pasta do código fonte
+cp $PROJECT_DIR"/config.py" $SCRIPT_DIR"/config.py"
+if [ "$?" != "0" ]; then
+    echo "Error copying $PROJECT_DIR/config.py. Does the file exist? Check out the sample project on projects/"
+    exit 1
+fi
 
 # Remove pasta de saída local, se existir
 rm -r $OUTPUT_DIR
@@ -14,9 +29,6 @@ rm -r $OUTPUT_DIR
 # Getting the absolute paths out of the relative ones, needed for the python script
 OUTPUT_DIR=$ROOT_DIR"/"$OUTPUT_DIR
 SCRIPT_DIR=$ROOT_DIR"/"$SCRIPT_DIR
-
-echo 'Output directory: '$OUTPUT_DIR
-echo 'Python script directory: '$SCRIPT_DIR
 
 # Vai para pasta do projeto
 cd $SCRIPT_DIR
